Debugged basic functionality (except my events) page
