+++ src/data_access/DataAccessInterface.java
@@ -4,6 +4,7 @@ import data_access.in_memory_dao.InMemoryCourseDataAccessObject;
 import data_access.in_memory_dao.InMemoryEmployeeDataAccessObject;
 import data_access.in_memory_dao.InMemoryEventDataAccessObject;
 import data_access.in_memory_dao.InMemorySessionDataAccessObject;
+import entity.Employee;
 
 public interface DataAccessInterface {
     public InMemoryCourseDataAccessObject getCourseDAO();
@@ -11,6 +12,9 @@ public interface DataAccessInterface {
     public InMemoryEventDataAccessObject getEventDAO();
     public InMemorySessionDataAccessObject getSessionDAO();
 
-    public boolean saveToDatabase();
+    public boolean saveToDatabase(InMemoryCourseDataAccessObject courseDAO,
+                                  InMemoryEmployeeDataAccessObject employeeDAO,
+                                  InMemoryEventDataAccessObject eventDAO,
+                                  InMemorySessionDataAccessObject sessionDAO);
 
 }
