+++ src/interface_adapter/get_events/getEventController.java
@@ -1,0 +1,4 @@
+package interface_adapter.get_events;
+
+public class getEventController {
+}
