+++ src/view/course_views/CreateCourseView.java
@@ -1,56 +1,64 @@
-package view.CourseViews;
+package view.course_views;
 
 import interface_adapter.UserState;
-import interface_adapter.enroll.EnrollViewModel;
+import interface_adapter.create_course.CreateCourseController;
+import interface_adapter.create_course.CreateCourseState;
+import interface_adapter.create_course.CreateCourseViewModel;
 import interface_adapter.ViewManagerModel;
-import interface_adapter.enroll.EnrollController;
-import interface_adapter.enroll.EnrollState;
 
+import javax.swing.*;
+import javax.swing.border.Border;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
-
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 
-import javax.swing.*;
-import javax.swing.border.Border;
+public class CreateCourseView extends JPanel implements ActionListener, PropertyChangeListener {
+    public final String viewName = "create course";
 
-// What is mySessionsViewName for?
-public class EnrollView extends JPanel implements ActionListener, PropertyChangeListener {
-    public final String viewName = "Invite to Course";
-    
     // Functionality variables
-    private EnrollViewModel enrollViewModel;
+    private CreateCourseViewModel createCourseViewModel;
     private ViewManagerModel viewManagerModel;
-    private EnrollController enrollController;
+    private CreateCourseController createCourseController;
     private UserState curUserState;
-    
+
 
     // Variables for setting up UI
-    private JButton invite;
-    private JButton close;
-    private final String HEADING_TEXT = "Invite To Course";
+    // Text
+    private final String HEADING_TEXT = "Create Course";
     private final String INPUT_FIELD_LABEL = "";
     private String error_message = "";
-    private JTextField userToEnrollInputField = new JTextField(10);
-    private JTextField courseCodeInputField = new JTextField(10);
-    private JLabel errorDisplayField = new JLabel(error_message);
-    
-    // Variables for Linking to other views
-    private String myCoursesViewName;
-    
-    
 
-    public EnrollView(EnrollController enrollController, EnrollViewModel enrollViewModel,
+    // Buttons
+    private JButton createNewCourse;
+    private JButton close;
+
+    // Input fields
+    private JTextField courseCodeTextField = new JTextField(10);
+    private JTextField courseNameTextField = new JTextField(10);
+    private JLabel errorDisplayField = new JLabel(error_message);
+
+    // Link to other views
+    private String myCoursesViewName;
+
+    /**
+     * Initializer for the little panel that pops up when you click make course in the my courses view.
+     * @param createCourseController Controller that takes in user input to get some output data
+     * @param createCourseViewModel The View Model that this pulls information from
+     * @param viewManagerModel Manages the current view to display
+     */
+    public CreateCourseView(CreateCourseController createCourseController, CreateCourseViewModel createCourseViewModel,
                       ViewManagerModel viewManagerModel, UserState curUserState){
-        this.enrollController = enrollController;
-        this.enrollViewModel = enrollViewModel;
+        this.createCourseController = createCourseController;
+        this.createCourseViewModel = createCourseViewModel;
         this.viewManagerModel = viewManagerModel;
+        createCourseViewModel.addPropertyChangeListener(this);
+
         this.curUserState = curUserState;
-        enrollViewModel.addPropertyChangeListener(this);
+
         setupUI();
     }
 
@@ -61,7 +69,7 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
         // Make components of UI
         JLabel headingLabel = makeHeading();
         JPanel buttonPanel = setUpButtons();
-        formatInputField();
+        formatInputFields();
         setUpErrorDisplayField();
 
         // Add components to UI
@@ -69,7 +77,7 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
         mainPanel.setLayout(new BoxLayout(mainPanel, BoxLayout.Y_AXIS));
         mainPanel.add(headingLabel);
 
-        JPanel inputFieldPanel = setUpInputFields();
+        JPanel inputFieldPanel = makeInputFields();
         mainPanel.add(inputFieldPanel);
 
         mainPanel.add(errorDisplayField);
@@ -91,19 +99,19 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
         return heading;
     }
 
-    private JPanel setUpInputFields() {
+    private JPanel makeInputFields() {
         JPanel inputFieldPanel = new JPanel();
         inputFieldPanel.setLayout(new BoxLayout(inputFieldPanel, BoxLayout.Y_AXIS));
 
         JPanel courseCodePanel = new JPanel();
         courseCodePanel.add(new JLabel("Course Code:"));
-        courseCodePanel.add(courseCodeInputField);
+        courseCodePanel.add(courseCodeTextField);
         inputFieldPanel.add(courseCodePanel);
 
-        JPanel userToEnrollPanel = new JPanel();
-        userToEnrollPanel.add(new JLabel("User to Enroll:"));
-        userToEnrollPanel.add(userToEnrollInputField);
-        inputFieldPanel.add(userToEnrollPanel);
+        JPanel courseNamePanel = new JPanel();
+        courseNamePanel.add(new JLabel("Course Name:"));
+        courseNamePanel.add(courseNameTextField);
+        inputFieldPanel.add(courseNamePanel);
 
         return inputFieldPanel;
     }
@@ -111,11 +119,11 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
     private JPanel setUpButtons() {
         JPanel buttonPanel = new JPanel();
 
-        invite = new JButton(enrollViewModel.ENROLL_BUTTON_LABEL);
-        invite.addActionListener(this);
-        buttonPanel.add(invite);
+        createNewCourse = new JButton(createCourseViewModel.CREATE_COURSE_BUTTON_LABEL);
+        createNewCourse.addActionListener(this);
+        buttonPanel.add(createNewCourse);
 
-        close = new JButton(enrollViewModel.CLOSE_BUTTON_LABEL);
+        close = new JButton(createCourseViewModel.CLOSE_BUTTON_LABEL);
         close.addActionListener(this);
         buttonPanel.add(close);
 
@@ -126,13 +134,14 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
         errorDisplayField.setAlignmentX(Component.CENTER_ALIGNMENT);
     }
 
-    private void formatInputField() {
+    private void formatInputFields() {
         setUpInputFieldFunctionality();
         setUpInputFieldStyle();
     }
 
     private void setUpInputFieldFunctionality() {
-        userToEnrollInputField.addKeyListener(new KeyListener() {
+        CreateCourseState currentState = createCourseViewModel.getState();
+        courseCodeTextField.addKeyListener(new KeyListener() {
             @Override
             public void keyTyped(KeyEvent e) {
             }
@@ -142,13 +151,12 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
 
             @Override
             public void keyReleased(KeyEvent e) {
-                EnrollState currentState = enrollViewModel.getState();
-                currentState.setUserInvited(userToEnrollInputField.getText());
-                enrollViewModel.setState(currentState);
+                currentState.setCourseCreated(courseCodeTextField.getText());
+                createCourseViewModel.setState(currentState);
             }
         });
 
-        courseCodeInputField.addKeyListener(new KeyListener() {
+        courseNameTextField.addKeyListener(new KeyListener() {
             @Override
             public void keyTyped(KeyEvent e) {
             }
@@ -158,32 +166,32 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
 
             @Override
             public void keyReleased(KeyEvent e) {
-                EnrollState currentState = enrollViewModel.getState();
-                currentState.setCourseCode(courseCodeInputField.getText());
-                enrollViewModel.setState(currentState);
+                currentState.setCourseName(courseNameTextField.getText());
+                createCourseViewModel.setState(currentState);
             }
         });
     }
 
     private void setUpInputFieldStyle() {
         Border paddingBorder = BorderFactory.createEmptyBorder(5, 5, 5, 5);
-        userToEnrollInputField.setBorder(BorderFactory.createCompoundBorder(
-                userToEnrollInputField.getBorder(), paddingBorder));
+        courseCodeTextField.setBorder(BorderFactory.createCompoundBorder(
+                courseCodeTextField.getBorder(), paddingBorder));
     }
 
     @Override
     public void actionPerformed(ActionEvent event) {
-        // When the user clicks the invite button, call a controller and pass in the employee to enroll
-        if (event.getSource() == invite) {
-            EnrollState curState = enrollViewModel.getState();
+        // When the user clicks the create new course button, call the controller and pass in the courses to create
+        if (event.getSource() == createNewCourse) {
 
-            String inviteeID = curState.getUserInvited();
-            String courseCode = curState.getCourseCode();
-            String curUserID = curUserState.getUserID();
+            CreateCourseState curState = createCourseViewModel.getState();
 
-            enrollController.enroll(curUserID, inviteeID, courseCode);
+            String courseCode = curState.getCourseCreated();
+            String courseName = curState.getCourseName();
+            String instructor = curUserState.getUserID();
 
-        // Close the window and return to the dashboard
+            createCourseController.createCourse(courseCode, courseName, instructor);
+
+        // Close the window and return to the myCourses view
         } else if (event.getSource() == close) {
             viewManagerModel.setActiveView(myCoursesViewName);
             viewManagerModel.firePropertyChanged();
@@ -197,11 +205,11 @@ public class EnrollView extends JPanel implements ActionListener, PropertyChange
     @Override
     public void propertyChange(PropertyChangeEvent event) {
         // Display the response message
-        EnrollState curState = enrollViewModel.getState();
-        errorDisplayField.setText(curState.getInviteResponseMessage());
+        CreateCourseState curState = createCourseViewModel.getState();
+        errorDisplayField.setText(curState.getCourseCreationResponseMessage());
         errorDisplayField.setForeground(Color.BLUE);
     }
-    
+
     public void linkViews(String myCoursesViewName) {
         this.myCoursesViewName = myCoursesViewName;
     }
