+++ src/view/course_views/EnrollView.java
@@ -1,65 +1,56 @@
-package view.CourseViews;
+package view.course_views;
 
-import entity.User;
 import interface_adapter.UserState;
-import interface_adapter.create_course.CreateCourseController;
-import interface_adapter.create_course.CreateCourseState;
-import interface_adapter.create_course.CreateCourseViewModel;
+import interface_adapter.enroll.EnrollViewModel;
 import interface_adapter.ViewManagerModel;
+import interface_adapter.enroll.EnrollController;
+import interface_adapter.enroll.EnrollState;
 
-import javax.swing.*;
-import javax.swing.border.Border;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
+
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 
-public class CreateCourseView extends JPanel implements ActionListener, PropertyChangeListener {
-    public final String viewName = "create course";
+import javax.swing.*;
+import javax.swing.border.Border;
 
+// What is mySessionsViewName for?
+public class EnrollView extends JPanel implements ActionListener, PropertyChangeListener {
+    public final String viewName = "Invite to Course";
+    
     // Functionality variables
-    private CreateCourseViewModel createCourseViewModel;
+    private EnrollViewModel enrollViewModel;
     private ViewManagerModel viewManagerModel;
-    private CreateCourseController createCourseController;
+    private EnrollController enrollController;
     private UserState curUserState;
-
+    
 
     // Variables for setting up UI
-    // Text
-    private final String HEADING_TEXT = "Create Course";
+    private JButton invite;
+    private JButton close;
+    private final String HEADING_TEXT = "Invite To Course";
     private final String INPUT_FIELD_LABEL = "";
     private String error_message = "";
-
-    // Buttons
-    private JButton createNewCourse;
-    private JButton close;
-
-    // Input fields
-    private JTextField courseCodeTextField = new JTextField(10);
-    private JTextField courseNameTextField = new JTextField(10);
+    private JTextField userToEnrollInputField = new JTextField(10);
+    private JTextField courseCodeInputField = new JTextField(10);
     private JLabel errorDisplayField = new JLabel(error_message);
-
-    // Link to other views
+    
+    // Variables for Linking to other views
     private String myCoursesViewName;
+    
+    
 
-    /**
-     * Initializer for the little panel that pops up when you click make course in the my courses view.
-     * @param createCourseController Controller that takes in user input to get some output data
-     * @param createCourseViewModel The View Model that this pulls information from
-     * @param viewManagerModel Manages the current view to display
-     */
-    public CreateCourseView(CreateCourseController createCourseController, CreateCourseViewModel createCourseViewModel,
+    public EnrollView(EnrollController enrollController, EnrollViewModel enrollViewModel,
                       ViewManagerModel viewManagerModel, UserState curUserState){
-        this.createCourseController = createCourseController;
-        this.createCourseViewModel = createCourseViewModel;
+        this.enrollController = enrollController;
+        this.enrollViewModel = enrollViewModel;
         this.viewManagerModel = viewManagerModel;
-        createCourseViewModel.addPropertyChangeListener(this);
-
         this.curUserState = curUserState;
-
+        enrollViewModel.addPropertyChangeListener(this);
         setupUI();
     }
 
@@ -70,7 +61,7 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
         // Make components of UI
         JLabel headingLabel = makeHeading();
         JPanel buttonPanel = setUpButtons();
-        formatInputFields();
+        formatInputField();
         setUpErrorDisplayField();
 
         // Add components to UI
@@ -78,7 +69,7 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
         mainPanel.setLayout(new BoxLayout(mainPanel, BoxLayout.Y_AXIS));
         mainPanel.add(headingLabel);
 
-        JPanel inputFieldPanel = makeInputFields();
+        JPanel inputFieldPanel = setUpInputFields();
         mainPanel.add(inputFieldPanel);
 
         mainPanel.add(errorDisplayField);
@@ -100,19 +91,19 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
         return heading;
     }
 
-    private JPanel makeInputFields() {
+    private JPanel setUpInputFields() {
         JPanel inputFieldPanel = new JPanel();
         inputFieldPanel.setLayout(new BoxLayout(inputFieldPanel, BoxLayout.Y_AXIS));
 
         JPanel courseCodePanel = new JPanel();
         courseCodePanel.add(new JLabel("Course Code:"));
-        courseCodePanel.add(courseCodeTextField);
+        courseCodePanel.add(courseCodeInputField);
         inputFieldPanel.add(courseCodePanel);
 
-        JPanel courseNamePanel = new JPanel();
-        courseNamePanel.add(new JLabel("Course Name:"));
-        courseNamePanel.add(courseNameTextField);
-        inputFieldPanel.add(courseNamePanel);
+        JPanel userToEnrollPanel = new JPanel();
+        userToEnrollPanel.add(new JLabel("User to Enroll:"));
+        userToEnrollPanel.add(userToEnrollInputField);
+        inputFieldPanel.add(userToEnrollPanel);
 
         return inputFieldPanel;
     }
@@ -120,11 +111,11 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
     private JPanel setUpButtons() {
         JPanel buttonPanel = new JPanel();
 
-        createNewCourse = new JButton(createCourseViewModel.CREATE_COURSE_BUTTON_LABEL);
-        createNewCourse.addActionListener(this);
-        buttonPanel.add(createNewCourse);
+        invite = new JButton(enrollViewModel.ENROLL_BUTTON_LABEL);
+        invite.addActionListener(this);
+        buttonPanel.add(invite);
 
-        close = new JButton(createCourseViewModel.CLOSE_BUTTON_LABEL);
+        close = new JButton(enrollViewModel.CLOSE_BUTTON_LABEL);
         close.addActionListener(this);
         buttonPanel.add(close);
 
@@ -135,14 +126,13 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
         errorDisplayField.setAlignmentX(Component.CENTER_ALIGNMENT);
     }
 
-    private void formatInputFields() {
+    private void formatInputField() {
         setUpInputFieldFunctionality();
         setUpInputFieldStyle();
     }
 
     private void setUpInputFieldFunctionality() {
-        CreateCourseState currentState = createCourseViewModel.getState();
-        courseCodeTextField.addKeyListener(new KeyListener() {
+        userToEnrollInputField.addKeyListener(new KeyListener() {
             @Override
             public void keyTyped(KeyEvent e) {
             }
@@ -152,12 +142,13 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
 
             @Override
             public void keyReleased(KeyEvent e) {
-                currentState.setCourseCreated(courseCodeTextField.getText());
-                createCourseViewModel.setState(currentState);
+                EnrollState currentState = enrollViewModel.getState();
+                currentState.setUserInvited(userToEnrollInputField.getText());
+                enrollViewModel.setState(currentState);
             }
         });
 
-        courseNameTextField.addKeyListener(new KeyListener() {
+        courseCodeInputField.addKeyListener(new KeyListener() {
             @Override
             public void keyTyped(KeyEvent e) {
             }
@@ -167,32 +158,32 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
 
             @Override
             public void keyReleased(KeyEvent e) {
-                currentState.setCourseName(courseNameTextField.getText());
-                createCourseViewModel.setState(currentState);
+                EnrollState currentState = enrollViewModel.getState();
+                currentState.setCourseCode(courseCodeInputField.getText());
+                enrollViewModel.setState(currentState);
             }
         });
     }
 
     private void setUpInputFieldStyle() {
         Border paddingBorder = BorderFactory.createEmptyBorder(5, 5, 5, 5);
-        courseCodeTextField.setBorder(BorderFactory.createCompoundBorder(
-                courseCodeTextField.getBorder(), paddingBorder));
+        userToEnrollInputField.setBorder(BorderFactory.createCompoundBorder(
+                userToEnrollInputField.getBorder(), paddingBorder));
     }
 
     @Override
     public void actionPerformed(ActionEvent event) {
-        // When the user clicks the create new course button, call the controller and pass in the courses to create
-        if (event.getSource() == createNewCourse) {
+        // When the user clicks the invite button, call a controller and pass in the employee to enroll
+        if (event.getSource() == invite) {
+            EnrollState curState = enrollViewModel.getState();
 
-            CreateCourseState curState = createCourseViewModel.getState();
+            String inviteeID = curState.getUserInvited();
+            String courseCode = curState.getCourseCode();
+            String curUserID = curUserState.getUserID();
 
-            String courseCode = curState.getCourseCreated();
-            String courseName = curState.getCourseName();
-            String instructor = curUserState.getUserID();
+            enrollController.enroll(curUserID, inviteeID, courseCode);
 
-            createCourseController.createCourse(courseCode, courseName, instructor);
-
-        // Close the window and return to the myCourses view
+        // Close the window and return to the dashboard
         } else if (event.getSource() == close) {
             viewManagerModel.setActiveView(myCoursesViewName);
             viewManagerModel.firePropertyChanged();
@@ -206,11 +197,11 @@ public class CreateCourseView extends JPanel implements ActionListener, Property
     @Override
     public void propertyChange(PropertyChangeEvent event) {
         // Display the response message
-        CreateCourseState curState = createCourseViewModel.getState();
-        errorDisplayField.setText(curState.getCourseCreationResponseMessage());
+        EnrollState curState = enrollViewModel.getState();
+        errorDisplayField.setText(curState.getInviteResponseMessage());
         errorDisplayField.setForeground(Color.BLUE);
     }
-
+    
     public void linkViews(String myCoursesViewName) {
         this.myCoursesViewName = myCoursesViewName;
     }
