+++ src/view/event_views/MyEventsViewInstructor.java
@@ -1,4 +1,4 @@
-package view.EventViews;
+package view.event_views;
 
 import interface_adapter.ViewManagerModel;
