+++ src/view/event_views/MyEventsViewTA.java
@@ -1,4 +1,4 @@
-package view.EventViews;
+package view.event_views;
 
 import interface_adapter.ViewManagerModel;
