+++ src/view/intro_views/SignUpView.java
@@ -1,4 +1,4 @@
-package view.IntroViews;
+package view.intro_views;
 
 import interface_adapter.ViewManagerModel;
 import interface_adapter.sign_up.SignUpController;
